feat: update curator complex predicate experiments for unified API
This commit updates the curator experiment scripts to work with the new unified
complex predicate API and parameter naming:

## curator.py changes:
- Added bitmap building timing measurement
- Integrated build_filter_bitmap() calls for all filters during setup
- Added bitmap_build_time_s metric to build results

## curator_with_index.py changes:
- Updated parameter names from nprobe/prune_thres to search_ef/beam_size/variance_boost
- Replaced index_bitmap_filter() with simplified index_filter() API
- Removed manual compute_qualified_labels() calls (handled internally)
- Updated parameter sweep functions to use new parameter spaces
- Simplified query calls by removing use_filter_labels parameter

## Key improvements:
- Consistent parameter naming across all curator experiments
- Unified filter indexing API reduces code complexity
- Better performance tracking with bitmap build timing
- Modernized parameter spaces aligned with latest curator capabilities

These changes ensure curator experiments work seamlessly with the updated
index wrappers and provide consistent benchmarking across all baseline methods.

+++ benchmark/complex_predicate/utils.py
@@ -177,4 +177,4 @@ def compute_qualified_labels(filter_str: str, train_mds: list[list[int]]) -> np.
     for i, md in enumerate(train_mds):
         if md and evaluate_predicate(filter_str, md):
             qualified_ids.append(i)
-    return np.array(qualified_ids, dtype=np.int64)
+    return np.array(qualified_ids, dtype=np.uint32)
